+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - **FEAT**: Add `controller` to `ShadSelect` and form fields, to control the selected values.
 - **FIX**: Improve the `ShadResizable` controller handling and simplify the logic to resize the panels.
 - **FIX**: `ShadResizable` handle position with `Axis.vertical`
+- **BREAKING CHANGE**: Now `ShadResizable` requires an `id`, to be able to handle when a panel is removed/added from the widget tree correctly.
 
 ## 0.18.7